Incorporate WS support
