+++ publisher/Cargo.toml
@@ -1,7 +1,12 @@
 [package]
 name = "publisher"
 version = "0.1.0"
+resolver = "2"
 edition = "2021"
+
+[features]
+ws = []
+sse = []
 
 [dependencies]
 async-stream = "0.3.5"
